+++ server/leaderboard_cache.go
@@ -579,8 +579,8 @@ func (l *LocalLeaderboardCache) ListTournaments(now int64, categoryStart, catego
 			// if (endTime == 0 && leaderboard.EndTime != 0) || (endTime == -1 && endTime < now) ||leaderboard.EndTime > endTime || leaderboard.EndTime == 0) || leaderboard.EndTime > endTime {
 			// SKIP tournaments where:
 			// - If end time filter is == 0, tournament end time is non-0.
-			// - If end time filter is to be ignored (-1) and tournament has ended.
-			// - If end time filter is set to a value and tournament end time is below it.
+			// - If end time filter is default (show only ongoing/future tournaments) and tournament has ended.
+			// - If end time filter is set and tournament end time is below it.
 			continue
 		}