+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project are documented below.
 The format is based on [keep a changelog](http://keepachangelog.com) and this project uses [semantic versioning](http://semver.org).
 
 ## [Unreleased]
-
+### Fixed
+- Better handling of SSL connections in development configurations.
 
 ## [2.14.1] - 2020-11-02
 ### Added