+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 - Make metrics prefix configurable, default to fixed value.
 - Build with Go 1.15.5 release.
 - Skip logging Lua errors raised by explicit runtime calls to the `error({msg, code})` built-in.
+- Update to Facebook Graph API v9.0.
+- Facebook authentication no longer requires access to gender, locale, and timezone data.
 
 ### Fixed
 - Better handling of SSL connections in development configurations.